requirements3.txt
-f /usr/share/pip-wheels
alabaster==0.7.11
alembic==1.0.0
aniso8601==3.0.2
appdirs==1.4.3
arrow==0.12.1
asn1crypto==0.24.0
astropy==3.0.3
async==0.6.2
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
backports.functools-lru-cache==1.5
banal==0.3.7
bcrypt==3.1.4
beautifulsoup4==4.6.0
biopython==1.72
black==18.6b4
bleach==2.1.3
blinker==1.4
bokeh==0.13.0
boto==2.49.0
boto3==1.7.83
botocore==1.10.83
bottle==0.12.13
Bottleneck==1.2.1
bottlenose==1.1.8
BTrees==4.5.0
bz2file==0.98
cachetools==2.1.0
cairocffi==0.8.1
CairoSVG==2.1.3
certifi==2018.8.13
cffi==1.11.5
cftime==1.0.1
Chameleon==3.4
chardet==3.0.4
cheroot==6.3.3
CherryPy==17.0.0
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cloudpickle==0.5.3
colorama==0.3.9
configobj==5.0.6
coverage==4.5.1
cryptography==2.2.2
cssselect==1.0.3
cssselect2==0.2.1
cycler==0.10.0
Cython==0.28.4
dask==0.18.1
dataset==1.1.0
decorator==4.3.0
dectate==0.13
defusedxml==0.5.0
Django==2.1
django-blog-zinnia==0.20
django-bootstrap-form==3.4
django-classy-tags==0.8.0
django-contrib-comments==1.8.0
django-formtools==2.1
django-js-asset==1.1.0
django-mptt==0.9.0
django-object-tools==1.11.0
django-sekizai==0.10.0
django-tagging==0.4.6
django-treebeard==4.3
django-xmlrpc==0.1.8
djangocms-admin-style==1.2.8
docopt==0.6.2
docutils==0.14
dominate==2.3.1
EasyProcess==0.2.3
entrypoints==0.2.3
et-xmlfile==1.0.1
falcon==1.4.1
feedfinder2==0.0.4
feedgenerator==1.9
feedparser==5.2.1
filebrowser-safe==0.5.0
Fiona==1.7.13
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Babel==0.11.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
future==0.16.0
Genshi==0.7
gensim==3.5.0
geographiclib==1.49
geopy==1.15.0
gevent==1.3.5
gitdb==0.6.4
gitdb2==2.0.4
GitPython==2.1.11
gmpy==1.17
google-api-core==1.16.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
google-cloud-vision==0.41.0
googleapis-common-protos==1.51.0
grappelli-safe==0.5.0
greenlet==0.4.13
grokcore.component==3.1
grpcio==1.26.0
h5py==2.8.0
html5lib==1.0.1
httplib2==0.11.3
hupper==1.3
idna==2.7
imagesize==1.0.0
IMAPClient==2.0.0
importscan==0.1
inflection==0.3.1
ipykernel==4.8.2
ipyparallel==5.0.1
ipython==4.1.2
ipython-genutils==0.2.0
ipywidgets==5.2.3
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jaraco.functools==1.20
jdcal==1.4
jellyfish==0.6.1
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
lxml==4.2.3
Mako==1.0.7
marisa-trie==0.7.5
Markdown==2.6.11
MarkupSafe==1.0
martian==1.2
matplotlib==2.2.2
mechanize==0.2.5
mimerender==0.6.0
mistune==0.8.3
mock==2.0.0
more-itertools==4.3.0
morepath==0.18.1
mots-vides==2015.5.11
mpmath==1.0.0
munch==2.3.2
mysql-connector-python==8.0.11
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
neo4j-driver==1.6.1
neotime==1.0.0
netCDF4==1.4.0
netifaces==0.10.7
networkx==2.1
newspaper3k==0.2.6
nltk==3.3
normality==0.6.1
nose==1.3.7
notebook==4.2.2
numexpr==2.6.5
numpy==1.14.5
oauth2client==4.1.2
oauthlib==2.1.0
opencv-contrib-python-headless==3.4.2.17
openpyxl==2.5.4
packaging==17.1
pandas==0.23.3
pandocfilters==1.4.2
paramiko==2.4.1
Paste==2.0.3
PasteDeploy==1.5.2
pbr==4.1.0
pdfrw==0.4
peewee==3.5.2
pelican==3.7.1
pep8==1.7.1
persistent==4.2.4.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
Pint==0.8.1
plaster==1.0
plaster-pastedeploy==0.6
plotly==3.0.0
pluggy==0.7.1
pocketsphinx==0.1.15
portend==2.3
praw==6.0.0
prawcore==1.0.0
prettytable==0.7.2
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.6
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pudb==2018.1
py==1.5.4
py-bcrypt==0.4
py2neo==4.0.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycairo==1.17.1
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyDatalog==0.17.1
pydub==0.22.1
pyenchant==2.0.0
pyflakes==2.0.0
pyflakes3k==0.4.3
pygal==2.4.0
pygeoip==0.3.2
Pygments==2.2.0
PyGObject==3.28.2
PyJWT==1.6.4
Pykka==1.2.1
pymc==2.3.6
pymongo==3.7.1
PyNaCl==1.2.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
Pyphen==0.9.4
pyquery==1.4.0
pyramid==1.9.2
pyserial==3.4
PySocks==1.6.8
pyspotify==2.0.5
PyStemmer==1.3.0
pytesseract==0.2.2
pytest==3.7.1
python-amazon-simple-product-api==2.1.0
python-dateutil==2.7.3
python-editor==1.0.3
python-gettext==3.0
python-Levenshtein==0.12.0
python-magic==0.4.15
python-mimeparse==1.6.0
python-slugify==1.2.5
pytz==2018.5
PyVirtualDisplay==0.2.1
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.0
qrcode==6.0
qtconsole==4.4.1
Quandl==3.4.0
rdflib==4.2.2
redis==2.10.6
reg==0.11
regex==2018.8.17
repoze.lru==0.7
requests==2.19.1
requests-cache==0.4.13
requests-file==1.4.3
requests-oauthlib==1.0.0
retrying==1.3.3
rpy2==2.8.6
rsa==3.4.2
ruffus==2.7.0
s3transfer==0.1.13
scikit-image==0.14.0
scipy==1.1.0
selenium==2.53.6
setproctitle==1.1.10
Shapely==1.6.4.post1
simplegeneric==0.8.1
simplejson==3.16.0
simpy==3.0.11
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.6.0
smmap==0.9.0
smmap2==2.0.4
snowballstemmer==1.2.1
socketpool==0.5.3
sockjs-tornado==1.0.3
SocksiPy-branch==1.1
South==1.0.2
SPARQLWrapper==1.8.2
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
splinter==0.7.3
SQLAlchemy==1.2.10
stevedore==1.28.0
stripe==2.0.1
sympy==1.2
TA-Lib==0.4.17
tables==3.4.4
Tempita==0.5.2
tempora==1.13
terminado==0.8.1
testpath==0.3.1
texcaller==0
textblob==0.15.1
tinycss==0.4
tinycss2==0.6.1
tldextract==2.2.0
toml==0.9.4
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
transaction==2.2.1
translationstring==1.3
twilio==6.15.0
twitter==1.18.0
tzlocal==1.5.1
uncertainties==3.0.2
Unidecode==1.0.22
update-checker==0.16
uritemplate==3.0.0
urllib3==1.23
urwid==2.0.1
venusian==1.1.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
visitor==0.1.3
w3lib==1.19.0
waitress==1.1.0
Wand==0.4.4
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5.1
WebOb==1.8.2
WebTest==2.0.30
Werkzeug==0.14.1
wheezy.caching==0.1.114
wheezy.core==0.1.140
wheezy.html==0.1.147
wheezy.http==0.1.344
wheezy.routing==0.1.157
wheezy.security==0.1.64
wheezy.template==0.1.167
wheezy.validation==0.1.135
wheezy.web==0.1.485
Whoosh==2.7.4
widgetsnbextension==1.2.6
WSGIProxy2==0.4.4
WTForms==2.2.1
xgboost==0.72.1
xlrd==1.1.0
xlwt==1.3.0
z3c.pt==3.1.0
zc.buildout==2.12.1
zc.lockfile==1.3.0
ZConfig==3.2.0
zdaemon==4.2.0
ZEO==5.2.0
ZODB==5.4.0
ZODB3==3.11.0
zodbpickle==1.0.2
zope.annotation==4.6.0
zope.app.appsetup==4.0.0
zope.app.publication==4.2.1
zope.app.wsgi==4.1.0
zope.authentication==4.3
zope.browser==2.2.0
zope.browserpage==4.2.0
zope.browserresource==4.2.1
zope.cachedescriptors==4.3.1
zope.catalog==4.2.1
zope.component==4.4.1
zope.configuration==4.1.0
zope.container==4.2.1
zope.contentprovider==4.1.0
zope.contenttype==4.3.0
zope.datetime==4.2.0
zope.deprecation==4.3.0
zope.dottedname==4.2
zope.error==4.4.0
zope.event==4.3.0
zope.exceptions==4.2.0
zope.filerepresentation==4.2.0
zope.formlib==4.4
zope.i18n==4.3.1
zope.i18nmessageid==4.1.0
zope.index==4.3.0
zope.interface==4.5.0
zope.intid==4.3.0
zope.keyreference==4.1.0
zope.lifecycleevent==4.2.0
zope.location==4.1.0
zope.login==2.1.0
zope.minmax==2.2.0
zope.pagetemplate==4.3.0
zope.password==4.3.1
zope.principalregistry==4.2.0
zope.processlifetime==2.2.0
zope.proxy==4.3.0
zope.ptresource==4.1.0
zope.publisher==4.3.2
zope.schema==4.5.0
zope.security==4.2.2
zope.securitypolicy==4.2.0
zope.session==4.2.0
zope.site==4.1.0
zope.size==4.2.0
zope.tal==4.3.1
zope.tales==4.2.0
zope.testbrowser==5.2.4
zope.testing==4.6.2
zope.traversing==4.2.0
zope.viewlet==4.1.0
